requirements.txt
streamlit
requests
pillow
pytesseract
nltk
textblob
python-dotenv
